tests/wpt/tests/css/css-masking/parsing/mask-size-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-size with invalid values</title>
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-size">
<meta name="assert" content="mask-size supports only the grammar '<bg-size>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-size", "-1px");
test_invalid_value("mask-size", "2% -3%");
test_invalid_value("mask-size", "1px 2px 3px");
</script>
</body>
</html>
